Still some problems with server but now it can send messages
