+++ main.py
@@ -15,7 +15,7 @@ from assets.Chat import Ui_PartisanMain
 from assets.AddContact import Ui_DialogAddContact
 from PyQt5 import QtWidgets
 
-from modules.server_queue import Queue
+from modules.server_queue import Queue, Server
 
 
 class ChatWindow(QtWidgets.QMainWindow, Ui_PartisanMain):
@@ -44,7 +44,7 @@ class ChatWindow(QtWidgets.QMainWindow, Ui_PartisanMain):
         self.pushButtonConfig.clicked.connect(self.contact_info)
 
     def listen(self):
-        self.server = ServerThread(self, self.account.ip, self.account.port)
+        self.server = Server(self.account.ip, self.account.port, self)
         self.server.start_server()
         self.actionConnection.setText("Connection: OK")
 
@@ -77,7 +77,6 @@ class ChatWindow(QtWidgets.QMainWindow, Ui_PartisanMain):
                                                                "JSON File (*.json)",
                                                                options=options)
         if self.accountJsonPath:
-            # self.setWindowTitle(f"Partisan - Messenger - {self.accountJsonPath}")
             self.centralwidget.setEnabled(True)
             with open(self.accountJsonPath, 'r') as account_file:
                 card = json.load(account_file)
@@ -138,6 +137,8 @@ class ChatWindow(QtWidgets.QMainWindow, Ui_PartisanMain):
                 print("Nice")
 
     def active_dialog(self):
+        self.pushButtonSendMessage.setDisabled(True)
+        self.lineInputMessage.clear()
         self.check_connection_thread.terminate()
         self.listMessages.clear()
         self.labelChat.setText("Chat")
@@ -216,37 +217,6 @@ class EditContactDialog(QtWidgets.QDialog, Ui_DialogContactInfo):
         self.close()
 
 
-class ServerThread(QThread):
-    def __init__(self, main_window, ip, port):
-        super().__init__()
-        self.ip = ip
-        self.port = port
-        self.main_window = main_window
-        self.queue = Queue(ip, port)
-
-    def start_server(self):
-        self.queue.start_server()
-
-    def stop_server(self):
-        self.queue.stop_server()
-
-    def loop(self):
-        while True:
-            time.sleep(1)
-            while self.queue.exists():
-                self.handle(self.queue.get())
-
-    def handle(self, message):
-        try:
-            # TODO: Func for incoming messages
-
-            print("Got: {}".format(message))
-            print(self)
-
-        except Exception as e:
-            print("Error: {}".format(e))
-
-
 class CheckConnectionThread(QThread):
     def __init__(self, main_window):
         super().__init__()
@@ -255,10 +225,10 @@ class CheckConnectionThread(QThread):
 
     def run(self):
         self.main_window.labelChatStatus.setText("Connection...")
-        print(self.contact_ip)
-        host = ping(self.contact_ip, count=5, interval=0.1, privileged=False)
+        host = ping(self.contact_ip, count=1, interval=0.1, privileged=False)
         if host.is_alive:
             self.main_window.labelChatStatus.setText("Online")
+            self.main_window.pushButtonSendMessage.setEnabled(True)
         else:
             self.main_window.labelChatStatus.setText("Offline")
         self.quit()
