+++ send_script.py
@@ -1,9 +1,9 @@
 import socket
 
 package = "Test"
-sock = socket.socket(socket.AF_INET6, socket.SOCK_STREAM)
+sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
-sock.connect(("localhost", 41030))
+sock.connect(('localhost', 41030))
 sock.send(str(package).encode('utf-8'))
 
 data = sock.recv(4096)
